[REVERTME] Baylake: Support USB installer
BZ: 86101

Installer supporting syslinux partition and sparse. It can be run on baylake
board from a USB stick.
Allow selection of prebuilt kernel or kernel built from sources.

This is required for baylake bring up as long as there is no solution
but to use the USB installer for initial setup.

Change-Id: Ib5f9cc71202a17a6641bb96360760ea46ecede99
Signed-off-by: Brian Wood <brian.j.wood@intel.com>
Signed-off-by: Guilhem IMBERTON <guilhem.imberton@intel.com>
Signed-off-by: Laurent FERT <laurent.fert@intel.com>
Reviewed-on: http://android.intel.com:8080/91404
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Smith, Geoff <geoff.smith@intel.com>
Tested-by: Ng, Cheon-woei <cheon-woei.ng@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>